move test out of manager
